deleted all the __init__.py dont know why we need them? Also continued work on server & client
